+++ ChangeLog.md
@@ -11,6 +11,7 @@
 - Added `ArticleIDs.mark_unread`.
   ([#10](https://github.com/davep/oldas/pull/10))
 - Added `Subscriptions.add`. ([#17](https://github.com/davep/oldas/pull/17))
+- Added `Subscriptions.remove`.
 
 ## v0.3.1