+++ ChangeLog.md
@@ -14,7 +14,7 @@
 - Added `Subscriptions.remove`.
   ([#18](https://github.com/davep/oldas/pull/18))
 - Added `Folders.rename`. ([#19](https://github.com/davep/oldas/pull/19))
-- Added `Folders.remove`.
+- Added `Folders.remove`. ([#20](https://github.com/davep/oldas/pull/20))
 
 ## v0.3.1