+++ ChangeLog.md
@@ -17,6 +17,8 @@
 - Added `Folders.remove`. ([#20](https://github.com/davep/oldas/pull/20))
 - Added `Subscriptions.rename`.
   ([#21](https://github.com/davep/oldas/pull/21))
+- Added `Subscriptions.move`.
+  ([#22](https://github.com/davep/oldas/pull/22))
 
 ## v0.3.1